Refactor Guice creation for the set-up of the prometheus exporter endpoint
