+++ override-api-web.xml
@@ -21,7 +21,7 @@
       <filter-name>prometheusFilter</filter-name>
       <url-pattern>/*</url-pattern>
     </filter-mapping>
- 
+
     <filter>
         <filter-name>cross-origin</filter-name>
         <filter-class>org.eclipse.jetty.servlets.CrossOriginFilter</filter-class>
