+++ src/main/java/uk/ac/cam/cl/dtg/segue/api/monitors/SegueMetrics.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.dtg.segue.api.metrics;
+package uk.ac.cam.cl.dtg.segue.api.monitors;
 
 import io.prometheus.client.Gauge;
 import io.prometheus.client.guava.cache.CacheMetricsCollector;
@@ -18,17 +18,6 @@ public final class SegueMetrics {
     public static final Gauge CURRENT_WEBSOCKET_USERS = Gauge.build()
             .name("segue_current_websocket_users").help("Currently number of websocket users/browsers.").register();
     // Anonymous user stats are calculated using metrics on the guava cache which holds a reference to each active user
-
-    // Streams Apps Metrics
-    public static final Gauge SITE_STATISTICS_STREAMS_APP_STATUS = Gauge.build()
-            .name("segue_site_statistics_streams_app_status")
-            .help("Site statistics Kafka streams app status (0: down 1: up)").register();
-    public static final Gauge USER_STATISTICS_STREAMS_APP_STATUS = Gauge.build()
-            .name("segue_user_statistics_streams_app_status")
-            .help("User statistics Kafka streams app status (0: down, 1: up)").register();
-    public static final Gauge ANONYMOUS_LOGGED_EVENTS_STREAMS_APP_STATUS = Gauge.build()
-            .name("segue_anonymous_logged_events_streams_app_status")
-            .help("Anonymous logged events Kafka streams app status (0: down, 1: up)").register();
 
     /**
      *  Private constructor as it does not make sense to instantiate this class.
